Remove text that only appears with dark mode plugin
There was a 'ghost goose' piece of text that only appeared when you enabled a specific dark mode plugin. The issue is that different dark mode plugins yield different results, and there is no clear preference on which one we should be using. As such, I'm removing this issue for now.