Reinstate answers.md file and add link to answers on landing page
There is a link to the previous answers page location on the GDS blog post about this repo. As it was a markdown file, I can't do a redirect from within the site. For now, I've added the page back ain and added a link to the new page on it.

I've also included a link to the answers on the home page, for anyone who might want to skip straight to them.